modularized frontend routes, updated folder structure
